requirements.txt
Django~=4.1.4
djangorestframework~=3.14.0
Pillow~=9.3.0
django-nested-admin~=4.0.2